requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pillow==10.4.0
numpy==2.0.1
requests==2.32.3
python-dotenv==1.0.1
scipy==1.13.1
